fix errors found by vet
Vet found the following errors:
	drivers/spi/mcp3002.go:74: (rx[0] & 0x3) (8 bits) too small for shift of 8
	drivers/spi/mcp3004.go:75: (rx[1] & 0x3) (8 bits) too small for shift of 8
	drivers/spi/mcp3008.go:75: (rx[1] & 0x3) (8 bits) too small for shift of 8
	drivers/spi/mcp3202.go:75: (rx[1] & 0xf) (8 bits) too small for shift of 8
	drivers/spi/mcp3204.go:75: (rx[1] & 0xf) (8 bits) too small for shift of 8
	drivers/spi/mcp3208.go:75: (rx[1] & 0xf) (8 bits) too small for shift of 8
	drivers/spi/mcp3304.go:75: (rx[1] & 0xf) (8 bits) too small for shift of 8
	platforms/dexter/gopigo3/driver.go:476: duty (8 bits) too small for shift of 8
	platforms/dexter/gopigo3/driver.go:556: response[5] (8 bits) too small for shift of 8

These are all about shifting bytes into larger values, but doing it in the wrong
order so the bits are shifted out completely. The fix is to do the upsize before
the shift.

The ones in spi are all errors of grouping, easily fixed by parenthesizing.

The first one in driver.go is not addressed here. It seems to be a plain but
harmless error: duty is an 8-bit value so shifting it right by 8 bits will always
yield 0.

The second expression can be greatly simplified as well as fixed. There is
no need to stage through a uint64. Go does not have C's sign conversion
craziness.