platforms/dexter/gopigo3: fix errors found by vet
Vet found the following errors:
	platforms/dexter/gopigo3/driver.go:491: duty (8 bits) too small for shift of 8
	platforms/dexter/gopigo3/driver.go:571: response[5] (8 bits) too small for shift of 8

The first error is caused by multiplying a byte by 100, overflowing it, but expecting it
to preserve all the bits. We address this by introducing a new, larger variable bigDuty.

The second error can be simplified as well as fixed. It is caused by shifting before
converting to a larger value.