[FIX] stepperDriver Run not using stepper speed
I'm not sure if anyone using it like this? Like blasting the stepperDriver with full force for loop steps.
At least my stepper is making noises like a badly wounded rabbit and is not moving anywhere.
Using a delay in Run() like in Move() fixed that. I'm pretty sure it was intended in that way.